requirements.txt
mysqlclient
django
django-environ
djangorestframework
django-filter
djangorestframework-simplejwt